requirements.txt
Flask==1.1.2
google-cloud-datastore==1.9.0
google-auth==1.6.2
requests==2.21.0
bcrypt==3.1.7
gunicorn
